requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
passlib
bcrypt
pyjwt
python-jose
python-multipart
pymysql
cryptography